refactor(react-gui): split shared/ipcTypes.ts into shared/types/<domain>.ts
Cut the 896-line contract file into 13 domain slices (layout, uiPrefs,
fileDialog, recent, fileEvents, appPath, naviCtxMenu, textCtxMenu,
sceneCtxMenu, menuState, crash, renderWindow, clipboard) so that a change to
one IPC domain no longer touches the file every other domain imports.
ipcTypes.ts is a transitional barrel for this commit; the importers are
rewritten next.
